Move all pages into index.page.tsx
